// admin/src/api/generated/schemas/adminAdminMenu.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

/**
 * AdminMenu represents a menu item (admin_menu).
 */
export type AdminAdminMenu = {
  id?: string;
  pid?: string;
  level?: number;
  tree?: string;
  name?: string;
  path?: string;
  icon?: string;
  hide_in_menu?: boolean;
  hide_children_in_menu?: boolean;
  sort?: number;
  remark?: string;
  status?: number;
  updated_at?: string;
  created_at?: string;
  deleted_at?: string;
  target?: string;
  access?: string;
  key?: string;
  children?: AdminAdminMenu[];
};