// admin/src/api/generated/schemas/adminAdminRole.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */
import type { AdminAdminMenu } from './adminAdminMenu';
import type { AdminAdminPermission } from './adminAdminPermission';

/**
 * AdminRole represents a role (admin_role).
 */
export type AdminAdminRole = {
  id?: string;
  name?: string;
  code?: string;
  description?: string;
  status?: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
  menus?: AdminAdminMenu[];
  permissions?: AdminAdminPermission[];
};